Seperates tests for each entity with folders, moves company tests to company folder
