+++ JobApplicationTrackerAPI.Tests/CompanyTests/CompanyRepositoryTests.cs
@@ -5,7 +5,7 @@ using JobApplicationTrackerAPI.Repository;
 using JobApplicationTrackerAPI.Repository.Company;
 using Microsoft.EntityFrameworkCore;
 
-namespace JobApplicationTrackerAPI.Tests
+namespace JobApplicationTrackerAPI.Tests.CompanyTests
 {
     public class CompanyRepositoryTests : IDisposable
     {
